Refactor inventory quantity handling to support decimal values:
- Updated various models, requests, and services to change quantity fields from integer to float, allowing for fractional quantities in inventory operations.
- Modified validation rules in requests to accept numeric values with a minimum of 0.001 for quantities.
- Adjusted database migration to change quantity columns to decimal(15,4) for better precision in stock management.

These changes enhance the accuracy and flexibility of inventory management across the application.

+++ app/Http/Controllers/Api/Admin/Inventory/DamageReportController.php
@@ -203,7 +203,7 @@ class DamageReportController extends Controller
         $company = Company::findOrFail($report->company_id);
 
         foreach ($report->damageReportItems as $item) {
-            $quantity = (int) $item->quantity;
+            $quantity = (float) $item->quantity;
             if ($quantity <= 0) {
                 continue;
             }
